chore: Rename some graph -> networks
