+++ enderio-conduits-modded/src/main/java/com/enderio/modconduits/common/modules/refinedstorage/RSConduitNodeData.java
@@ -3,7 +3,6 @@ package com.enderio.modconduits.common.modules.refinedstorage;
 import com.enderio.conduits.api.network.node.IConduitNode;
 import com.enderio.conduits.api.network.node.NodeData;
 import com.enderio.conduits.api.network.node.NodeDataType;
-import com.mojang.serialization.MapCodec;
 import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridNetworkNode;
 import com.refinedmods.refinedstorage.api.network.node.NetworkNode;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
@@ -21,9 +20,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class RSConduitNodeData implements NodeData {
-    public static final MapCodec<RSConduitNodeData> CODEC = MapCodec.unit(RSConduitNodeData::new);
-
-    public static NodeDataType<RSConduitNodeData> TYPE = new NodeDataType<>(CODEC, RSConduitNodeData::new);
+    public static NodeDataType<RSConduitNodeData> TYPE = new NodeDataType<>(null, RSConduitNodeData::new);
 
     public final NetworkNodeContainerProvider container;
     private ConduitRSNode mainNode;
