+++ enderio-conduits/src/main/java/com/enderio/conduits/common/conduit/bundle/ConduitBundleBlockEntity.java
@@ -22,8 +22,8 @@ import com.enderio.conduits.common.conduit.legacy.ConnectionState;
 import com.enderio.conduits.common.conduit.legacy.DynamicConnectionState;
 import com.enderio.conduits.common.conduit.legacy.StaticConnectionStates;
 import com.enderio.conduits.common.conduit.menu.ConduitMenu;
-import com.enderio.conduits.common.conduit.new_graph.ConduitNetworkSavedData;
-import com.enderio.conduits.common.conduit.new_graph.ConduitNode;
+import com.enderio.conduits.common.conduit.network.ConduitNetworkSavedData;
+import com.enderio.conduits.common.conduit.network.ConduitNode;
 import com.enderio.conduits.common.init.ConduitBlockEntities;
 import com.enderio.conduits.common.init.ConduitTypes;
 import com.enderio.core.common.blockentity.EnderBlockEntity;
