+++ enderio-conduits/src/main/java/com/enderio/conduits/common/conduit/network/package-info.java
@@ -1,4 +1,4 @@
 @javax.annotation.ParametersAreNonnullByDefault
 @net.minecraft.MethodsReturnNonnullByDefault
 
-package com.enderio.conduits.common.conduit.new_graph;
+package com.enderio.conduits.common.conduit.network;
