+++ enderio-conduits/src/main/java/com/enderio/conduits/common/conduit/type/energy/EnergyConduit.java
@@ -45,7 +45,7 @@ public record EnergyConduit(ResourceLocation texture, Component description, int
 
     // Not configurable - energy is instantaneous
     @Override
-    public int graphTickRate() {
+    public int networkTickRate() {
         return 1;
     }
