+++ enderio-conduits/src/main/java/com/enderio/conduits/common/conduit/type/fluid/FluidConduit.java
@@ -225,7 +225,7 @@ public record FluidConduit(ResourceLocation texture, Component description, int 
 
         if (pTooltipFlag.hasShiftDown()) {
             String rawRateFormatted = String.format("%,d",
-                    (int) Math.ceil(transferRatePerTick() * (20.0 / graphTickRate())));
+                    (int) Math.ceil(transferRatePerTick() * (20.0 / networkTickRate())));
             pTooltipAdder.accept(TooltipUtil.styledWithArgs(ConduitLang.FLUID_RAW_RATE_TOOLTIP, rawRateFormatted));
         }
     }
