+++ enderio-conduits/src/main/java/com/enderio/conduits/common/conduit/type/redstone/RedstoneConduit.java
@@ -39,7 +39,7 @@ public record RedstoneConduit(ResourceLocation texture, ResourceLocation activeT
             .apply(builder, RedstoneConduit::new));
 
     @Override
-    public int graphTickRate() {
+    public int networkTickRate() {
         return 2;
     }
